Set initial project structure and configurations.
In this commit we define the basic structure of the project,
separating the models, services and api-routes. We also add
two linters, prettier and eslint. Using both linters is good
because while eslint scans our js files for common syntax and
style errors, prettier scans for style errors and automatically
correct them, ensuring that rules are being followed.

We set up the linters to run before tests in package.json file.